+++ CHANGELOG.md
@@ -6,8 +6,57 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## Unreleased
 ### Added
-- E-mail address revalidation, addressing issues where user's e-mail addresses can be (temporary) invalid
+- Option `skipExpiryCheck` in disclosure requests to allow disclosure of expired credentials (e.g. `"skipExpiryCheck": ["irma-demo.sidn-pbdf.email"]`)
+- Option `host` in session request to overrule host name in IRMA QR if permission has been granted (see below)
+  ```
+  {
+    "@context": "https://irma.app/ld/request/disclosure/v2",
+    "host": "irma.example.com",
+    "disclose": ...
+  }
+  ```
+  This leads to the following session package:
+  ```
+  {
+    "token":"KzxuWKwL5KGLKr4uerws",
+    "sessionPtr": {"u":"https://irma.example.com/irma/session/ysDohpoySavbHAUDjmpz","irmaqr":"disclosing"},
+    "frontendRequest": {
+      "authorization":"qGrMmL8UZwZ88Sq8gobV",
+      "minProtocolVersion": "1.0",
+      "maxProtocolVersion": "1.1"
+    }
+  }
+  ```
+- Permission option `host_perms` in the requestor configuration to specify which values a requestor may use for the `host` option in session requests
+  ```
+  {
+    "requestors": {
+        "myapp": {
+            "disclose_perms": [ "irma-demo.MijnOverheid.ageLower.over18" ],
+            "sign_perms": [ "irma-demo.MijnOverheid.ageLower.*" ],
+            "issue_perms": [ "irma-demo.MijnOverheid.ageLower" ],
+            "host_perms": ["*.example.com"]
+            "auth_method": "token",
+            "key": "eGE2PSomOT84amVVdTU"
+        }
+    }
+  }
+  ```
+- Renewal endpoint for keyshare attribute in the keyshare server (`/users/renewKeyshareAttribute`)
 - Keyshare server /api/v2/prove/... endpoints for the new keyshare protocol
+
+### Changed
+- `KeyshareVerifyPin` function in irmaclient ensures the keyshare attribute is valid
+- Sending the account expiry email is done when user has only valid e-mail addresses
+- Strip unnecessary details from database errors
+
+### Fixed
+- User account expiry continues when one or more e-mail addresses are marked for revalidation
+
+## [0.13.3] - 2023-09-06
+### Fixed
+- Auto-update mechanism of IRMA configuration not working in ghcr.io/privacybydesign/irma Docker container
+- Panics occur when the timestamp file does not exist in a scheme directory
 
 ## [0.13.2] - 2023-08-22
 ### Changed
@@ -406,6 +455,7 @@ This release contains several large new features. In particular, the shoulder su
 - Combined issuance-disclosure requests with two schemes one of which has a keyshare server now work as expected
 - Various other bugfixes
 
+[0.13.3]: https://github.com/privacybydesign/irmago/compare/v0.13.2...v0.13.3
 [0.13.2]: https://github.com/privacybydesign/irmago/compare/v0.13.1...v0.13.2
 [0.13.1]: https://github.com/privacybydesign/irmago/compare/v0.13.0...v0.13.1
 [0.13.0]: https://github.com/privacybydesign/irmago/compare/v0.12.6...v0.13.0